v0.29 Menu Improvement & Bug Fixes
The How To Play button in the main menu has been activated and a popup has been added where a simple pretutorial can be read.

The Credits button on the main menu replaced with the story button and added a popup where the story can be read.

Fixed the bug of earning multiple points in a single egg.